rename core input type
