+++ examples/transport-http_headers__dynamicHeaders.output.txt
@@ -1,10 +1,11 @@
 {
-  url: 'https://countries.trevorblades.com/graphql',
-  body: '{"query":"{ languages { code } }"}',
-  method: 'POST',
   headers: Headers {
     accept: 'application/graphql-response+json; charset=utf-8, application/json; charset=utf-8',
     'content-type': 'application/json',
-    'x-sent-at-time': '1725560229910'
-  }
+    'x-sent-at-time': '1725564397598'
+  },
+  signal: undefined,
+  method: 'POST',
+  url: 'https://countries.trevorblades.com/graphql',
+  body: '{"query":"{ languages { code } }"}'
 }
